# docs/swagger.yaml
definitions:
...
  dto.SyncOrdersRequest:
    properties:
      force_sync:
        description: 强制全量同步
        type: boolean
      max_created:
        type: string
      min_created:
        description: Unix timestamp
        type: string
      shop_id:
        type: integer
    required:
    - shop_id
    type: object
...
  dto.UpdateOrderNoteRequest:
    properties:
      message_from_seller:
        type: string
    type: object
  dto.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - pending
        - processing
        - shipped
        - delivered
        - canceled
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
paths:
...
  /api/orders:
    get:
      consumes:
      - application/json
      description: 分页查询订单，支持按店铺、状态、关键词、日期范围筛选
      parameters:
      - description: 店铺ID
        in: query
        name: shop_id
        type: integer
      - description: 订单状态
        in: query
        name: status
        type: string
      - description: 关键词搜索（买家名/订单号）
        in: query
        name: keyword
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.ListOrdersResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单列表
      tags:
      - Order (订单管理)
  /api/orders/{id}:
    get:
      description: 根据订单ID获取完整订单信息，包含订单项、收货地址、发货信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.OrderDetailResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 订单不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单详情
      tags:
      - Order (订单管理)
  /api/orders/{id}/note:
    patch:
      consumes:
      - application/json
      description: 更新卖家对订单的备注信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 备注内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "备注已更新"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新订单备注
      tags:
      - Order (订单管理)
  /api/orders/{id}/shipment:
    get:
      description: 获取指定订单的发货记录详情
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.ShipmentVO}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 发货服务未配置
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单发货信息
      tags:
      - Order (订单管理)
  /api/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: 更新订单的处理状态（如标记为已处理、已取消等）
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "状态已更新"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新订单状态
      tags:
      - Order (订单管理)
  /api/orders/{order_id}/shipment:
    get:
      description: 获取指定订单的发货记录
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.EtsyShipmentResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据订单ID获取发货信息
      tags:
      - Shipment (发货管理)
  /api/orders/stats:
    get:
      description: 获取订单统计数据（订单数、金额、状态分布等）
      parameters:
      - description: 店铺ID
        in: query
        name: shop_id
        type: integer
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.OrderStatsResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单统计
      tags:
      - Order (订单管理)
  /api/orders/sync:
    post:
      consumes:
      - application/json
      description: 从 Etsy 拉取订单数据并同步到本地数据库
      parameters:
      - description: 同步参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SyncOrdersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": {}, "message": "订单同步完成"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 同步失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 同步 Etsy 订单
      tags:
      - Order (订单管理)
...
swagger: "2.0"